+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.2.0] - 2018-12-10
+
 ### Added
 
 - [github.com/InVisionApp/go-logger](https://github.com/InVisionApp/go-logger) integration
@@ -24,4 +27,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial release
 
 
-[Unreleased]: https://github.com/goph/logur/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/goph/logur/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/goph/logur/compare/0.1.0...v0.2.0